--- Unidad 2-Estructuras de Control/Equipo 1/Ejercicios/Ejercicio 1 - Daniel/JS/problema1.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<head>
-  <meta charset="UTF-8">
-  <title>Test Ejercicio 1</title>
-  <script src="problema1.js"></script>
-</head>
-<body></body>
-
-  <!-- Para testear el programa abrir este archivo en cualquier navegador actual (Chrome, Safari, etc) -->
-  
-</body>
-</html>
