--- Unidad 2-Estructuras de Control/Equipo 1/Ejercicios/Ejercicio 13 - Eyder/C/ejercicio13C.depend
@@ -1,5 +1,0 @@
-# depslib dependency file v1.0
-1549325791 source:c:\users\eyder\desktop\tareas\programación estructuraada\programacionestructurada\unidad 2-estructuras de control\equipo 1\ejercicios\ejercicio13\ejercicio13c\main.c
-	<stdio.h>
-	<stdlib.h>
-
