--- Unidad 2-Estructuras de Control/Equipo 1/Ejercicios/Ejercicio 13 - Eyder/C/ejercicio13C.layout
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
-<CodeBlocks_layout_file>
-	<FileVersion major="1" minor="0" />
-	<ActiveTarget name="Debug" />
-	<File name="main.c" open="1" top="1" tabpos="1" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
-		<Cursor>
-			<Cursor1 position="475" topLine="27" />
-		</Cursor>
-	</File>
-</CodeBlocks_layout_file>
