--- Unidad 2-Estructuras de Control/Equipo 1/Ejercicios/Ejercicio 13 - Eyder/JS/testerDeJs.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<head>
-  <title>Test Ejercicio 13</title>
-  <script src="ejercicio13.js"></script>
-</head>
-<body></body>
-
-  <!-- Para testear el programa abrir este archivo en cualquier navegador actual (Chrome, Safari, etc) -->
-  
-</body>
-</html>
