--- Unidad 2-Estructuras de Control/Equipo 1/Ejercicios/Ejercicio 19 - Daniel/JS/problema19Tester.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<head>
-  <meta charset="UTF-8">
-  <title>Test Ejercicio 19</title>
-  <script src="problema19.js"></script>
-</head>
-<body></body>
-
-  <!-- Para testear el programa abrir este archivo en cualquier navegador actual (Chrome, Safari, etc) -->
-  
-</body>
-</html>
