--- Unidad 2-Estructuras de Control/Equipo 1/Ejercicios/Ejercicio 25 - Jimmy/JS/25.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<head>
-  <title>Test Ejercicio 25</title>
-  <script src="25.js"></script>
-</head>
-<body></body>
-
-  <!-- Para testear el programa abrir este archivo en cualquier navegador actual (Chrome, Safari, etc) -->
-
-</body>
-</html>
