--- Unidad 2-Estructuras de Control/Equipo 1/Ejercicios/Ejercicio 37 - Eyder/C/ejercicio37C.depend
@@ -1,5 +1,0 @@
-# depslib dependency file v1.0
-1549417193 source:c:\users\eyder\desktop\tareas\programación estructuraada\programacionestructurada\unidad 2-estructuras de control\equipo 1\ejercicios\ejercicio37\ejercicio37c\main.c
-	<stdio.h>
-	<stdlib.h>
-
