--- Unidad 2-Estructuras de Control/Equipo 1/Ejercicios/Ejercicio 43 - Jimmy/JS/43.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<head>
-  <title>Test Ejercicio 43</title>
-  <script src="43.js"></script>
-</head>
-<body></body>
-
-  <!-- Para testear el programa abrir este archivo en cualquier navegador actual (Chrome, Safari, etc) -->
-
-</body>
-</html>
