--- Unidad 2-Estructuras de Control/Equipo 1/Ejercicios/Ejercicio 48 - Eyder/C/ejercicio48C.depend
@@ -1,5 +1,0 @@
-# depslib dependency file v1.0
-1549425499 source:c:\users\eyder\desktop\tareas\programación estructuraada\programacionestructurada\unidad 2-estructuras de control\equipo 1\ejercicios\ejercicio48\ejercicio48c\main.c
-	<stdio.h>
-	<stdlib.h>
-
