--- Unidad 2-Estructuras de Control/Equipo 1/Ejercicios/Ejercicio 48 - Eyder/JS/testerDeJs48.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<head>
-  <meta charset="UTF-8">
-  <title>Test Ejercicio 48</title>
-  <script src="ejercicio48.js"></script>
-</head>
-<body></body>
-
-  <!-- Para testear el programa abrir este archivo en cualquier navegador actual (Chrome, Safari, etc) -->
-  
-</body>
-</html>
