--- Unidad 2-Estructuras de Control/Equipo 1/Ejercicios/Ejercicio 7 - Daniel/JS/problema7.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<head>
-  <meta charset="UTF-8">
-  <title>Test Ejercicio 7</title>
-  <script src="problema7.js"></script>
-</head>
-<body></body>
-
-  <!-- Para testear el programa abrir este archivo en cualquier navegador actual (Chrome, Safari, etc) -->
-  
-</body>
-</html>
