--- Unidad 2-Estructuras de Control/Equipo 1/readme.md
@@ -1,42 +1,0 @@
-# Equipo 1
----
-### Conformado por: <br>
-+ Daniel Eduardo Gutiérrez Delfín
-+ Jimmy Nathan Ojeda Arana 
-+ Eyder Antonio Concha Moreno
----
-
-### Reparto de Ejercicios:
-
-| Integrantes   | Ejercicios    |
-| ------------- |:-------------:|
-| Jimmy         | 25, 43, 31    |
-| Daniel        | 1, 19, 7      |
-| Eyder         | 13, 37, 48    |
-
----
-
-### Lista de Ejercicios:
-
-#### 1.- Escribe un programa que reciba un número y te diga si ese número es par o impar <br>
-#### 7.- Escribir el programa para un programa que reciba un numero positivo, si este numero es mayor a 1000 se le sumara un 5%, si el numero es mayor a 3000 se le sumara otro 10% adicional y si el numero es mayor a 5000 se le sumara otro 5% adicional. Notese que al final
- número > 1000 5% <br>
- número > 3000 5% + 10% <br>
- número > 5000 5% + 10% + 5% <br>
-#### 13.- Escribir el programa para convertir una hora leída en horario militar (número de cuatro dígitos) y la imprima a una hora en formato estándar (hh:00am/pm).<br>
-#### 19.- Escribir un programa que lea tres números y determine el mayor de los tres.
-#### 25.- Escribir un programa para mostrar los códigos ASCII de las letras mayúsculas y minúsculas.
-#### 31.- Dado un número determinar si es primo.
-#### 37.- Determinar la suma de los primeros n números primos.
-#### 43.- Escribe el programa que lea N valores y que cuente cuantos de ellos son negativos y cuantos son positivos (cero se condición de fin de lectura).
-#### 48.- Una compañía efectúa cálculos para su reparto de utilidades. En esta ocasión añadirá un pequeño incentivo de acuerdo a la antigüedad de sus empleados. Se sabe que la compañía lleva 15 años trabajando y el incentivo se dará de acuerdo al número de años laborados según la información siguiente: 
-
-| Antiguedad |             | % de sobresueldo |
-|:----------:|:-----------:|:----------------:|
-| De         | A           |                  |
-| 1          | 3           |        1%        |
-| 4          | 6           |        3%        |
-| 7          | 9           |        5%        |
-| 10         | En adelante |        7%        |
-
-#### Escribe el programa que lea el sueldo del empleado y su antigüedad y que calcule su incentivo, imprimiendo su sueldo base, el incentivo y el sueldo a pagar. Suponer que no se conoce el número de empleados.
